feat(appointment): read copilot description bc i forgot what i did exactly
