+++ frontend/src/app/dashboard/appointments/AppointmentDetail/Appointment-Detail.module.css
