+++ frontend/src/app/make-appointment/New-Appointment-Customer.module.css
@@ -1,0 +1,33 @@
+.layout {
+    display: grid;
+    height: 100%;
+}
+
+.header {
+    display: grid;
+    grid-auto-flow: column;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 2em;
+    background-color: rgb(33, 33, 33);
+    padding: 0.5em;
+}
+
+.header h1 {
+    font-size: 1.25rem;
+    color: white;
+}
+
+.header figure {
+    max-width: 100px;
+    justify-self: center;
+}
+
+.header img {
+    width: 100%;
+}
+
+.scrollableContent {
+    background-color: #f5f5f5;
+    overflow: auto;
+    padding: 1em 0.5em;
+}
