+++ frontend/src/app/make-appointment/page.jsx
@@ -5,7 +5,12 @@ import styles from "./New-Appointment-Customer.module.css";
 export default function NewAppointmentCustomer() {
   return (
     <div className={styles.layout}>
-      <h1>Reservar cita</h1>
+      <header className={styles.header}>
+        <h1>Reservar cita</h1>
+        <figure>
+          <img src="icons/Sagoz_Icon.png" alt="Sagoz Logo" />
+        </figure>
+      </header>
       <main className={styles.scrollableContent}>
         <AppointmentForm mode={"customer"}></AppointmentForm>
       </main>
