+++ frontend/src/app/my-appointments/[id]/Appointment-Detail-Customer.module.css
@@ -4,7 +4,7 @@
 }
 
 .layout>h1 {
-    font-size: 1.4rem;
+    font-size: 1.25rem;
     padding: 0.5em;
     background-color: rgb(33, 33, 33);
     color: white;
