+++ frontend/src/app/signin/Sign-In.module.css
@@ -38,6 +38,17 @@
     grid-template-rows: auto 1fr auto;
     max-width: 20em;
     width: 100%;
+    row-gap: 2em;
+}
+
+.main figure {
+    max-width: 250px;
+    justify-self: center;
+}
+
+.main img {
+    width: 100%;
+    object-fit: cover;
 }
 
 .main h1 {
