+++ frontend/src/app/utils/routes.js
@@ -1,16 +1,15 @@
-export const signInRoute = "/signin"
+export const signInRoute = "/signin";
 
 export const appointmentsRoute = "/dashboard/appointments";
-export const edit = '/${id}/update';
-export const see = '/${id}';
+export const edit = "/${id}/update";
+export const see = "/${id}";
 export const newAppointmentRoute = appointmentsRoute + "/new";
 export const seeAppointments = appointmentsRoute + see;
 export const editAppointments = appointmentsRoute + edit;
 
-
 export const servicesRoute = "/dashboard/services";
 export const newServiceRoute = servicesRoute + "/new";
-export const newBundleRoute = servicesRoute + "/bundles/new"
+export const newBundleRoute = servicesRoute + "/bundles/new";
 export const seeService = servicesRoute + see;
 export const editService = servicesRoute + edit;
 
@@ -18,3 +17,5 @@ export const staffRoute = "/dashboard/staff";
 export const newStaffRoute = staffRoute + "/new";
 export const editStaffRoute = staffRoute + edit;
 export const seeStaffRoute = staffRoute + see;
+
+export const makeAppointmentRoute = "/make-appointment";
