+++ frontend/src/middleware.js
@@ -1,39 +1,35 @@
-import { NextResponse } from 'next/server'
-import { deleteToken } from './app/utils/requestBuilder';
+import { NextResponse } from "next/server";
 
 export default function middleware(request) {
-  const token = request.cookies.get('token');
+  const token = request.cookies.get("token");
 
   if (!token) {
-    return NextResponse.redirect(new URL('/', request.url));
+    return NextResponse.redirect(new URL("/", request.url));
   }
 
   let payload;
-  try{
+  try {
     payload = JSON.parse(
-      Buffer.from(token.value.split('.')[1], 'base64').toString()
+      Buffer.from(token.value.split(".")[1], "base64").toString()
     );
-  }catch(err){
-    const response = NextResponse.redirect(new URL('/', request.url));
-    response.cookies.delete('token');
+  } catch (err) {
+    const response = NextResponse.redirect(new URL("/", request.url));
+    response.cookies.delete("token");
     return response;
   }
-  
+
   const now = Math.floor(Date.now() / 1000);
 
   if (payload.exp < now) {
-    const response = NextResponse.redirect(new URL('/', request.url));
-    response.cookies.delete('token');
+    const response = NextResponse.redirect(new URL("/", request.url));
+    response.cookies.delete("token");
     return response;
   }
-
-
-
 
   return NextResponse.next();
 }
 
 // Proteger solo estas rutas:
 export const config = {
-  matcher: ['/dashboard/:path*', '/account'],
-}
+  matcher: ["/dashboard/:path*", "/account"],
+};
